--- Decsys/ClientApp/src/components/ComponentEditor/ImageUpload.stories.js
@@ -1,7 +1,0 @@
-import React from "react";
-import { storiesOf } from "@storybook/react";
-import ImageUpload from "./ImageUpload";
-
-storiesOf("Admin/ComponentEditor/ImageUpload", module).add("Default", () => (
-  <ImageUpload params={{}} />
-));
