--- Decsys/ClientApp/src/components/ComponentEditor/Param.stories.js
@@ -1,22 +1,0 @@
-import React from "react";
-import { storiesOf } from "@storybook/react";
-import Param from "./Param";
-import { types } from "@decsys/param-types";
-import { text, boolean, optionsKnob } from "@storybook/addon-knobs";
-
-const oneOf = ["Hello", "Goodbye", "Something else"];
-
-storiesOf("Admin/ComponentEditor/Param", module)
-  .add("Text", () => (
-    <Param type={types.string} value={text("Value", "My Value")} />
-  ))
-  .add("Boolean", () => (
-    <Param type={types.bool} value={boolean("Value", true)} />
-  ))
-  .add("oneOf", () => (
-    <Param
-      type={types.oneOf}
-      value={optionsKnob("Value", oneOf, "Hello", { display: "inline-radio" })}
-      oneOf={oneOf}
-    />
-  ));
