+++ Decsys/ClientApp/src/stories/ComponentEditor/ComponentEditor.stories.js
@@ -1,0 +1,19 @@
+import React from "react";
+import ComponentEditor from "components/ComponentEditor";
+import {
+  paramsLookup,
+  knob,
+  components
+} from "components/ComponentRender/ComponentRender.stories";
+
+export default {
+  title: "Admin/ComponentEditor",
+  component: ComponentEditor
+};
+
+export const Basic = () => (
+  <ComponentEditor
+    component={components[knob()]}
+    params={paramsLookup[knob()].params}
+  />
+);
