+++ Decsys/ClientApp/src/stories/ComponentEditor/ImageUpload.stories.js
@@ -1,0 +1,9 @@
+import React from "react";
+import ImageUpload from "components/ComponentEditor/ImageUpload";
+
+export default {
+  title: "Admin/ComponentEditor/ImageUpload",
+  component: ImageUpload
+};
+
+export const Basic = () => <ImageUpload params={{}} />;
